move the request for editing the car to logic/editCarRequest.js file
