+++ src/containers/EditCar.js
@@ -1,37 +1,14 @@
 import PropTypes from 'prop-types';
 import { connect } from 'react-redux';
 import EditCarForm from '../components/EditCarForm';
-
-const sendEditCar = async (auth, cardata, carId) => {
-  try {
-    const body = {
-      mark: cardata.mark,
-      model: cardata.model,
-      year: cardata.year,
-      price: cardata.price,
-    };
-    const options = {
-      method: 'PUT',
-      headers: {
-        Accept: 'application/json',
-        'Content-Type': 'application/json',
-        Authorization: auth.authToken,
-      },
-      body: JSON.stringify(body),
-    };
-    await fetch(`http://[::1]:3000/cars/${carId}`, options);
-    return true;
-  } catch (error) {
-    return false;
-  }
-};
+import sendEditCarRequest from '../logic/editCarRequest';
 
 const EditCar = ({ location, auth }) => {
   const { car } = location.state;
 
   const handleSubmit = (cardata) => {
     if (auth.admin) {
-      sendEditCar(auth, cardata, car.id);
+      sendEditCarRequest(auth, cardata, car.id);
     }
   };
